+++ client/src/components/Navbar.jsx
@@ -42,11 +42,11 @@ function Navbar() {
             <div id="nav-right">
                 <button id="about-button" onClick={about}>About</button>
                 <div id="user-status">
-                    {/* {user ? ( */}
+                    {user ? (
                         <ProfileMenu/>
-                    {/* ) : (
+                    ) : (
                         <button id="get-started" onClick={getstarted}>Get Started</button>
-                    )} */}
+                    )}
                 </div>
             </div>
         </nav>