+++ server/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [1.0] - 8/5/25
+
+### Added
+- test route for login
+
+### Changed
+
+### Fixed
+
 ## [1.0] - 8/4/25
 
 ### Added